fix(crdt): widen Global m_size/m_capacity to uint32_t (regression from SBO)
Global::resize_zero doubles m_capacity until it covers the requested
size. For replica_id ≥ 32768 the doubling reaches 65536, which silently
truncates to 0 when stored back into the uint16_t m_capacity. The
corrupted Global then writes through m_inline (because on_heap() flips
to false) and propagates to copy/assign/join paths — repro is a heap-
buffer-overflow on the next operator= memcpy (caught under ASan), seen
intermittently as SIGSEGV in markoff-bench-render.

The same narrowing also bites observe(replica_id == 65535) where
ts.replica_id + 1 == 65536 truncated to 0 in resize_zero(uint16_t).

Reproducer (Markoff foundation-exploration handoff,
docs/handoff/2026-04-30-collabtext-sbo-regression-repro.md): a fresh
Buffer with replicaId in [4, 65535] crashes ~70 % of bench runs on the
first apply_local_edit. Existing tst_clock used replica_id ≤ 5, so the
SBO heap-promotion path was never exercised.

Fix: widen m_size/m_capacity (and the matching loop indices and
resize_zero param) to uint32_t. SBO_CAP and SBO inline storage are
unchanged. Three new regression tests in tst_clock cover the high
replica_id, the uint16_t boundary, and copy paths after promotion.

+++ libs/collabtext/src/crdt/Clock.cpp
@@ -66,7 +66,7 @@ Global::~Global() {
     if (on_heap()) std::free(m_heap_data);
 }
 
-void Global::resize_zero(uint16_t new_size) {
+void Global::resize_zero(uint32_t new_size) {
     if (new_size <= m_size) {
         m_size = new_size;
         return;
@@ -80,7 +80,7 @@ void Global::resize_zero(uint16_t new_size) {
         std::memset(new_data + m_size, 0, (new_cap - m_size) * sizeof(uint32_t));
         if (on_heap()) std::free(m_heap_data);
         m_heap_data = new_data;
-        m_capacity = static_cast<uint16_t>(new_cap);
+        m_capacity = new_cap;
     } else {
         std::memset(data_mut() + m_size, 0, (new_size - m_size) * sizeof(uint32_t));
     }
@@ -107,7 +107,7 @@ bool Global::observed_all(const Global &other) const {
     if (m_size < other.m_size) return false;
     const uint32_t *a = data();
     const uint32_t *b = other.data();
-    for (uint16_t i = 0; i < other.m_size; ++i)
+    for (uint32_t i = 0; i < other.m_size; ++i)
         if (a[i] < b[i]) return false;
     return true;
 }
@@ -117,15 +117,15 @@ void Global::join(const Global &other) {
     if (other.m_size > m_size) resize_zero(other.m_size);
     uint32_t *a = data_mut();
     const uint32_t *b = other.data();
-    for (uint16_t i = 0; i < other.m_size; ++i)
+    for (uint32_t i = 0; i < other.m_size; ++i)
         if (b[i] > a[i]) a[i] = b[i];
 }
 
 void Global::meet(const Global &other) {
-    uint16_t minLen = std::min(m_size, other.m_size);
+    uint32_t minLen = std::min(m_size, other.m_size);
     uint32_t *a = data_mut();
     const uint32_t *b = other.data();
-    for (uint16_t i = 0; i < minLen; ++i) {
+    for (uint32_t i = 0; i < minLen; ++i) {
         if (a[i] > 0 && b[i] > 0)
             a[i] = std::min(a[i], b[i]);
     }
@@ -137,7 +137,7 @@ bool Global::operator==(const Global &other) const {
     if (m_size != other.m_size) return false;
     const uint32_t *a = data();
     const uint32_t *b = other.data();
-    for (uint16_t i = 0; i < m_size; ++i)
+    for (uint32_t i = 0; i < m_size; ++i)
         if (a[i] != b[i]) return false;
     return true;
 }
